requirements.txt
streamlit
google-genai
supabase
pymarc
pandas
requests
pillow